% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{com.spdf}
\alias{com.spdf}
\title{Grand Paris Metropole Communes SpatialPolygonsDataFrame}
\description{
SpatialPolygonsDataFrame of the Grand Paris Metropole communes.\cr 
Fields: \cr
\itemize{
\item{DEPCOM: Commune identifiers}
}
}
\examples{
data(GrandParisMetropole)
sp::plot(com.spdf)
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3} for communes.
\url{http://www.apur.org/article/composition-12-territoires-metropole-grand-paris} 
for Grand Paris communal composition.
}

