% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{ept}
\alias{ept}
\title{Grand Paris Metropole EPTs}
\format{
\describe{
\item{EPT}{EPT identifiers}
\item{LIBEPT}{EPT names}
}
}
\source{
Atelier parisien d'urbanisme, Grand Paris communal composition (2015-12-17):\cr
\url{https://www.apur.org/fr/nos-travaux/composition-12-territoires-metropole-grand-paris}
}
\description{
sf object. Grand Paris Metropole EPTs. 
EPTs (Etablissements Publics Territoriaux) are groups of communes.\cr
}
\examples{
library(sf)
ept <- st_read(system.file("metroparis.gpkg", package = "MTA"), layer = "ept", quiet = TRUE)
}
