% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{ept.spdf}
\alias{ept.spdf}
\title{Grand Paris Metropole EPTs SpatialPolygonsDataFrame}
\description{
SpatialPolygonsDataFrame Grand Paris Metropole EPTs. 
EPTs (etablissements publics territoriaux) are groups of communes.\cr
Fields: \cr
\itemize{
\item{EPT: EPT identifiers}
\item{LIBEPT: EPT names}
}
}
\examples{
data(GrandParisMetropole)
sp::plot(ept.spdf)
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3} for communes.
\url{http://www.apur.org/article/composition-12-territoires-metropole-grand-paris} 
for Grand Paris communal composition.
}

