% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MTA.R
\name{globalDev}
\alias{globalDev}
\title{Global deviation}
\usage{
globalDev(spdf, id, stock1, stock2)
}
\arguments{
\item{spdf}{Spatial polygon data frame}

\item{id}{Unique identifier of the data frame (character)}

\item{stock1}{Numerator}

\item{stock2}{Denominator}
}
\value{
dataframe
}
\description{
This function compute the global deviation from two count data according to the MyperAtlas Methodolgy.
}
\details{
The general context may be the whole chosen study area. In such a case, the associated map will be the same as the associated map to the ratio itself. So, the user may choose another general context or a reference value. For instance, in the example of the EU, even if the study area is the 29 potential countries, it may be of interest to observe the spatial differentiations according to another global reference, for instance the global value associated to EU15. For this level, the user may also exogenously enter a value. By default, this value has first been set to the value of the global area.
}
\examples{
dev <-globalDev(spdf=nuts2006,id="nuts3",stock1="gdp2008",stock2="popt2008")
head(dev)
}

