% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA.R
\name{localDev}
\alias{localDev}
\title{Local deviation}
\usage{
localDev(spdf, x, spdfid = NULL, xid = NULL, var1, var2, order = NULL,
  dist = NULL, mat = NULL, type = "rel")
}
\arguments{
\item{spdf}{a SpatialPolygonsDataFrame.}

\item{x}{a data frame.}

\item{spdfid}{identifier field in spdf, default to the first column 
of the spdf data frame. (optional)}

\item{xid}{identifier field in x, default to the first column 
of x (optional)}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{order}{contiguity order.}

\item{dist}{distance defining the contiguity.}

\item{mat}{use an extern distance matrix (road km, time...).}

\item{type}{type of deviation; "rel" for relative deviation, "abs" for 
absolute deviation.}
}
\value{
A vector is returned.
}
\description{
This function compute the local deviation from two count 
data according to the MyperAtlas Methodolgy.
}
\details{
The spatial context shows which proximity relation will be 
the basis of the neighborhood’s definition for each elementary unit. 
That is usually “contiguity”, but it may also be a relationship based 
on distances since they have been introduced in the hyp file 
(units that are less than X kilometers far from), or time-distances. 
Then, each elementary unit value will be compared to the value of its 
neighborhood.
}

