% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta_get_data.R
\name{mf_get_mtq}
\alias{mf_get_mtq}
\alias{mta_get_data}
\title{Get the 'mta' dataset}
\usage{
mta_get_data()
}
\value{
two sf objects on Metropole du Grand Paris (communes and EPCI)
}
\description{
Import the mta example dataset (municipalities and EPCI of the Metropole du Grand Paris area).
}
\details{
This a wrapper around
}
\examples{
mta_get_data()
}
