\name{MTA-package}
\alias{MTA-package}
\docType{package}
\title{
Longitudinal microbiome data analysis: microbial trend analysis, group comparison and classification
}
\description{
Microbial trend analysis (MTA) framework is an attractive and effective tool in investigating the dynamic microbial pattern in the longitudinal microbiome study. In particular, MTA can perform three tasks: 1) MTA can capture the common and major community level microbial dynamic trends for a given group of subjects and identify the dominant taxa; 2) MTA can examine whether the microbial overall dynamic trends are significantly different or not in groups; 3) MTA can classify a given subject based on its longitudinal microbial profiling. Both simulations and real data applications showed that MTA framework is robust and powerful in hypothesis testing, taxon identification, and subject classification. }
\details{
\tabular{ll}{
Package: \tab MTA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-12-24\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Chan Wang, Jiyuan Hu, Martin J. Blaser, Huilin Li.

Maintainer: Chan Wang <Chan.Wang@nyulangone.org>, Huilin Li <Huilin.Li@nyulangone.org>
}
\references{
Wang C, Hu J, Blaser M J, Li H (2019). Longitudinal microbiome data analysis: microbial trend analysis, group comparison and classification.

}
\keyword{package}
