% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MTA.R
\name{nuts2006}
\alias{nuts2006}
\title{Spatial dataframe (polygons) of the NUTS regions (2006)}
\description{
European regions (NUTS 2006 version)
}
\section{Fields}{

\describe{
\item{\code{nuts3}}{nuts3 code}

\item{\code{nuts2}}{nuts2 code}

\item{\code{nuts1}}{nuts1 code}

\item{\code{nuts0}}{nuts0 code}

\item{\code{gdp2008}}{Gross domestic product in 2008 (euros) - source : ESPON database, 2014}

\item{\code{gdppps2008}}{Gross domestic product (ppp) in 2008 (euros) - source : ESPON database, 2014}

\item{\code{popt2008}}{total population in 2008 (inh.) - source : ESPON database, 2014}
}}

