% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Realdata.R
\docType{data}
\name{Realdata}
\alias{Realdata}
\title{Real data used in the Real data analysis section}
\format{
A list with the length=3 with names Control, STAT, and phy.tree:
\describe{
  \item{Control}{A 20 * 35 * 4 Array of relative abundances for 20 Controls}
  \item{STAT}{A 17 * 35 * 4 Array of relative abundances for 17 STAT}
  \item{phy.tree}{The phylogenetic tree (phylo-class) object for 35 taxa}
}
}
\usage{
Realdata
}
\description{
Realdata is a list with the length=3.
}
\keyword{datasets}
