% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA.R
\name{synthesis}
\alias{synthesis}
\title{Multiscalar typology}
\usage{
synthesis(spdf, id, stock1, stock2, distance = 0, key, threshold = 100,
  superior = TRUE, map = FALSE)
}
\arguments{
\item{spdf}{Spatial polygon data frame}

\item{id}{Unique identifier of the data frame (character)}

\item{stock1}{Numerator}

\item{stock2}{Denominator}

\item{key}{Aggregation ky for the medium deviatiopn}

\item{threshold}{Threshold of typology}

\item{superior}{If true, the criterium is "gretter than". If false, the criterium is "lower than"}

\item{map}{If true, the typology is ploted}

\item{dist}{distance (in the units of the basemap) between centroids to determinate the local context. 0 = contiguity.}
}
\value{
dataframe
}
\description{
This function compute a multiscalar typology according to the three relative positions about contexts (global, medium, local). The elementary units are classified in eight classes according to their three relative positions.
}
\details{
xxx
}
\examples{
# synthesis (without plot)
typology <-synthesis(spdf=nuts2006,id="nuts3",stock1="gdp2008",stock2="popt2008",distance=100000,key="nuts0",threshold=125,superior=TRUE,map=FALSE)

# synthesis (with plot)
typology <-synthesis(spdf=nuts2006,id="nuts3",stock1="gdp2008",stock2="popt2008",distance=100000,key="nuts0",threshold=125,superior=TRUE,map=TRUE)

}

