% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{synthesis3}
\alias{synthesis3}
\title{Multiscalar typology}
\usage{
synthesis3(spdf, x, spdfid = NULL, xid = NULL, var1, var2, ref = NULL,
  key, order = NULL, dist = NULL, mat = NULL, threshold,
  superior = FALSE)
}
\arguments{
\item{spdf}{Spatial polygon data frame.}

\item{x}{a dataframe.}

\item{spdfid}{identifier field in spdf, default to the first column 
of the spdf data frame (optional).}

\item{xid}{identifier field in x, default to the first column 
of x (optional).}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{ref}{distance (in the units of the basemap) between centroids to determinate the local context. 0 = contiguity.}

\item{key}{Aggregation key for the territorial deviation.}

\item{order}{contiguity order (dist and mat are not used).}

\item{dist}{distance defining the contiguity (order is not used).}

\item{mat}{a distance matrix (road distance, travel time...). Row and column 
names must fit xid identifiers.}

\item{threshold}{defined to build the typology (100 is considered as the average).}

\item{superior}{deviation values must be greater than threshold when criterion is true,
it must be lower than threshold when criterion is false.}
}
\value{
A dataframe including the ratio, the 3 relative deviations, their position according to the 
threshold and superior parameters, the resulting typology 8 classes and a proposal of colours for displaying territorial 
units on maps is returned.
}
\description{
This function compute a multiscalar typology according to the three relative 
deviations (global, territorial, spatial). The elementary units are classified
in eight classes according to their three relative positions.
}
\examples{
data("GrandParisMetropole")
synthesis <- synthesis3(spdf = com.spdf,
                        x = com,
                        spdfid = "DEPCOM",
                        xid = "DEPCOM",
                        var1 = "INC",
                        var2 = "TH",
                        dist = NULL,
                        key = "EPT",
                        order = 1,
                        mat = NULL,
                        threshold = 125,
                        superior = TRUE)

if(require('cartography')){
}
}

