% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesisAbs.R
\name{synthesisAbs}
\alias{synthesisAbs}
\title{Multiscalar absolute synthesis}
\usage{
synthesisAbs(spdf, x, spdfid = NULL, xid = NULL, var1, var2, ref = NULL,
  key, order = NULL, dist = NULL, mat = NULL)
}
\arguments{
\item{spdf}{Spatial polygon data frame.}

\item{x}{a dataframe.}

\item{spdfid}{identifier field in spdf, default to the first column 
of the spdf data frame (optional).}

\item{xid}{identifier field in x, default to the first column 
of x (optional).}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{ref}{distance (in the units of the basemap) between centroids to determinate the local context. 0 = contiguity.}

\item{key}{Aggregation key for the territorial deviation.}

\item{order}{contiguity order (dist and mat are not used).}

\item{dist}{distance defining the contiguity (order is not used).}

\item{mat}{a distance matrix (road distance, travel time...). Row and column 
names must fit xid identifiers.}
}
\value{
A dataframe including the mass of numerator to redistribue to reach a perfect equilibrium 
according to the 3 contexts, expressed in numerator measure unit and as a share of the numerator mass.
}
\description{
This function sums the total amount of redistributions according to the three absolute 
deviations (global, territorial, spatial).
}
\examples{
data("GrandParisMetropole")
}

