% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{com.spdf}
\alias{com.spdf}
\title{Grand Paris Metropole Communes SpatialPolygonsDataFrame}
\format{\describe{
\item{DEPCOM}{Commune identifiers}
}}
\source{
Institut national de l’information géographique et forestière (IGN), GEOFLA® 
2015 v2.1 Communes France Métropolitaine: \cr
\url{http://professionnels.ign.fr/geofla}\cr
Atelier parisien d'urbanisme, Grand Paris communal composition (2015-12-17):\cr
\url{http://www.apur.org/article/composition-12-territoires-metropole-grand-paris}
}
\description{
SpatialPolygonsDataFrame of the Grand Paris Metropole communes.\cr
}
\examples{
data(GrandParisMetropole)
sp::plot(com.spdf)
}
