% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{ept.spdf}
\alias{ept.spdf}
\title{Grand Paris Metropole EPTs SpatialPolygonsDataFrame}
\format{\describe{
\item{EPT}{EPT identifiers}
\item{LIBEPT}{EPT names}
}}
\source{
Atelier parisien d'urbanisme, Grand Paris communal composition (2015-12-17):\cr
\url{http://www.apur.org/article/composition-12-territoires-metropole-grand-paris}
}
\description{
SpatialPolygonsDataFrame of Grand Paris Metropole EPTs. 
EPTs (Etablissements Publics Territoriaux) are groups of communes.\cr
}
\examples{
data(GrandParisMetropole)
sp::plot(ept.spdf)
}
