% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA.R
\name{globalDev}
\alias{globalDev}
\title{Global Deviation}
\usage{
globalDev(x, var1, var2, ref = NULL, type = "rel")
}
\arguments{
\item{x}{a data frame.}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{ref}{ratio of reference; if NULL the mean ratio.}

\item{type}{type of deviation; "rel" for relative deviation, "abs" for 
absolute deviation.}
}
\value{
A vector is returned
}
\description{
This function compute the global deviation from two 
count data according to the HyperAtlas Methodolgy (methodo), see details.
}
\details{
The general context may be the whole chosen study area. 
In such a case, the associated map will be the same as the associated 
map to the ratio itself. So, the user may choose another general context 
or a reference value. For instance, in the example of the EU, even if 
the study area is the 29 potential countries, it may be of interest 
to observe the spatial differentiations according to another global 
reference, for instance the global value associated to EU15. For this 
level, the user may also exogenously enter a value. By default, this value 
has first been set to the value of the global area.
}
\examples{
library(cartography)
x <- nuts3.df
x$gdppps2008 <- x$gdppps2008 * 1000000
x[substr(x = x$id, start = 1, stop = 2) == "DE", "pop2008"] <- NA
x$gdevabs <- globalDev(x = x, 
                       var1 = "gdppps2008", 
                       var2 = "pop2008", 
                       type = "abs")
x$gdevrel <- globalDev(x = x, 
                       var1 = "gdppps2008", 
                       var2 = "pop2008", 
                       type = "rel")
par(mar = c(0,0,1.1,0))
choroLayer(spdf = nuts3.spdf, df = x, var = "gdevrel", 
           legend.pos = "topright", 
           breaks = c(11,50,75,100,125,150,7000), 
           col = carto.pal(pal1 = "blue.pal", n1 = 3, 
                           pal2 = "wine.pal", n2 = 3))
propSymbolsLayer(spdf = nuts3.spdf, df = x, var = "gdevabs", 
                 legend.pos = "right",
                 col = "#ff000050",col2 = "#0000ff50", 
                 k = 0.001, legend.style = "e",
                 breakval = 0)
layoutLayer(title = "Global Deviation")
}

