% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalDev.R
\name{globalDev}
\alias{globalDev}
\title{Global Deviation}
\usage{
globalDev(x, var1, var2, ref = NULL, type = "rel")
}
\arguments{
\item{x}{a data frame.}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{ref}{ratio of reference; if NULL, the ratio of reference is the one of 
the whole study area (data.frame).}

\item{type}{type of deviation; "rel" for relative deviation, "abs" for 
absolute deviation.}
}
\value{
A vector is returned.
}
\description{
This function computes the deviation between regional ratios 
and a ratio of reference. 
Each elementary unit value will be compared to a global value.
}
\examples{
library(cartography)
x <- nuts3.df
x$gdppps2008 <- x$gdppps2008 * 1000000
x$gdevabs <- globalDev(x = x, 
                       var1 = "gdppps2008", 
                       var2 = "pop2008", 
                       type = "abs")
x$gdevrel <- globalDev(x = x, 
                       var1 = "gdppps2008", 
                       var2 = "pop2008", 
                       type = "rel")
par(mar = c(0,0,1.2,0))
choroLayer(spdf = nuts3.spdf, df = x, var = "gdevrel", 
           legend.pos = "topright", 
           breaks = c(11,50,75,100,125,150,7000), border = NA,
           col = carto.pal(pal1 = "blue.pal", n1 = 3, 
                           pal2 = "wine.pal", n2 = 3))
propSymbolsLayer(spdf = nuts3.spdf, df = x, var = "gdevabs", 
                 legend.pos = "right",legend.values.rnd = -5,
                 col = "#ff000050",col2 = "#0000ff50", 
                 legend.style = "c", inches = 0.1,
                 breakval = 0)
plot(rgeos::gUnaryUnion(nuts0.spdf), add = TRUE)
layoutLayer(title = "Global Deviation")
}

