% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MTA.R
\name{localDev}
\alias{localDev}
\title{Local deviation}
\usage{
localDev(spdf, id, stock1, stock2, distance)
}
\arguments{
\item{spdf}{Spatial polygon data frame}

\item{id}{Unique identifier of the data frame (character)}

\item{stock1}{Numerator}

\item{stock2}{Denominator}

\item{dist}{distance (in the units of the basemap) between centroids to determinate the local context. 0 = contiguity.}
}
\value{
dataframe
}
\description{
This function compute the medium deviation from two count data according to the MyperAtlas Methodolgy.
}
\details{
The spatial context shows which proximity relation will be the basis of the neighborhood’s definition for each elementary unit. That is usually “contiguity”, but it may also be a relationship based on distances since they have been introduced in the hyp file (units that are less than X kilometers far from), or time-distances. Then, each elementary unit value will be compared to the value of its neighborhood.
}
\examples{
dev <-localDev(spdf=nuts2006,id="nuts3",stock1="gdp2008",stock2="popt2008",distance=0)
head(dev)
}

