% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDev.R
\name{localDev}
\alias{localDev}
\title{Local Deviation}
\usage{
localDev(spdf, x, spdfid = NULL, xid = NULL, var1, var2, order = NULL,
  dist = NULL, mat = NULL, type = "rel")
}
\arguments{
\item{spdf}{a SpatialPolygonsDataFrame.}

\item{x}{a data frame.}

\item{spdfid}{identifier field in spdf, default to the first column 
of the spdf data frame (optional).}

\item{xid}{identifier field in x, default to the first column 
of x (optional)}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{order}{contiguity order (dist and mat are not used).}

\item{dist}{distance defining the contiguity (order is not used).}

\item{mat}{a distance matrix (road distance, travel time...). Row and column 
names must fit xid identifiers.}

\item{type}{type of deviation; "rel" for relative deviation, "abs" for 
absolute deviation.}
}
\value{
A vector is returned.
}
\description{
This function computes the deviation between regional ratios and 
local ratios. Local ratios are defined either by a contiguity order or by a 
distance measure between regions.
Each elementary unit value will be compared to the value of its 
neighborhood.
}

