% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MTA.R
\name{mediumDev}
\alias{mediumDev}
\title{Medium deviation}
\usage{
mediumDev(spdf, id, stock1, stock2, key)
}
\arguments{
\item{spdf}{Spatial polygon data frame}

\item{id}{Unique identifier of the data frame (character)}

\item{stock1}{Numerator}

\item{stock2}{Denominator}

\item{key}{Aggregation key}
}
\value{
dataframe
}
\description{
This function compute the medium deviation from two count data according to the MyperAtlas Methodolgy.
}
\details{
The territorial context, on the other hand, has to be a geographical zoning that is an aggregation of the “elementary zoning” that was previously chosen.
}
\examples{
dev <-mediumDev(spdf=nuts2006,id="nuts3",stock1="gdp2008",stock2="popt2008",key="nuts0")
head(dev)
}

