% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA.R
\name{mediumDev}
\alias{mediumDev}
\title{Medium Deviation}
\usage{
mediumDev(x, var1, var2, key, type = "rel")
}
\arguments{
\item{x}{a data frame.}

\item{var1}{name of the numerator variable in x.}

\item{var2}{name of the denominator variable in x.}

\item{key}{aggregation key field.}

\item{type}{type of deviation; "rel" for relative deviation, "abs" for 
absolute deviation.}
}
\value{
A vector is returned
}
\description{
This function compute the medium deviation from two count 
data according to the HyperAtlas Methodolgy.
}
\details{
The territorial context, on the other hand, has to be a 
geographical zoning that is an aggregation of the “elementary zoning” 
that was previously chosen.
}
\examples{
library(cartography)
x <- nuts3.df
x$gdppps2008 <- x$gdppps2008 * 1000000
x[substr(x = x$id, start = 1, stop = 2) == "BE", "pop2008"] <- NA
x$key <- substr(x = x$id, start = 1, stop = 3)
x$mdevabs <- mediumDev(x = x, var1 = "gdppps2008", var2 = "pop2008", 
                       type = "abs", key = "key")
x$mdevrel <- mediumDev(x = x, var1 = "gdppps2008", 
                       var2 = "pop2008", 
                       type = "rel", key = "key")
par(mar = c(0,0,0,0))
plot(nuts0.spdf[nuts0.spdf$id \%in\% c("BE", "NL", "LU", "DE", "FR"),] )
choroLayer(spdf = nuts3.spdf, df = x, var = "mdevrel", 
           legend.pos = "topright", breaks = c(11,50,75,100,125,150,7000), 
           col = carto.pal(pal1 = "blue.pal", 3, pal2 = "wine.pal", 3), 
           border = NA, add=TRUE)
plot(nuts1.spdf, add=TRUE)
propSymbolsLayer(spdf = nuts3.spdf, df = x, var = "mdevabs", 
                 col = "#ff000050",col2 = "#0000ff50", k = 0.001, 
                 breakval = 0)
}

