\name{mt.analysis-package}
\alias{mt.analysis-package}
\alias{mt.analysis}
\docType{package}
\title{
mt.analysis: process 2D motor-tracing data from 2-alternative choice tasks
}
\description{
Functions to preprocess (time-normalize & compute derivatives of trajectory) and aggregate trajectory
}
\details{
\tabular{ll}{
Package: \tab mt.analysis\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2015-04-07\cr
License: \tab ?\cr
}

}
\author{
Jonas Haslbeck

Maintainer: Jonas Haslbeck <jonashaslbeck@gmail.com>
}
