% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MTA}
\alias{MTA}
\alias{MTA-package}
\title{Multiscalar Territorial Analysis}
\description{
Build multiscalar territorial analysis based on various contexts. \cr
Main functions : 
\itemize{
\item{\code{\link{generalDev}}: deviation between regional ratios and a 
ratio of reference.}
\item{\code{\link{territorialDev}}: deviation between regional ratios and 
ratios of an aggregated level.}
\item{\code{\link{localDev}}}: deviation between regional ratios and ratios
of neigborhing regions.
}
}
\references{
RONNNNNAAAAAANNNNN, A Great Paper Made With Love,
\emph{and in English}, 2023
}

