% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MTA}
\alias{MTA}
\alias{MTA-package}
\title{Multiscalar Territorial Analysis}
\description{
Build multiscalar territorial analysis based on various contexts for a given ratio defined by a numerator and a denominator.\cr
Main functions : 
\itemize{
\item{\code{\link{gdev}}: general deviation of each territorial unit as regards  
to all the study area (or a reference value).}
\item{\code{\link{tdev}}: territorial deviation of each territorial unit as regards  
to an intermediate territorial level of reference.}
\item{\code{\link{sdev}}}: spatial deviation of each territorial unit as regards  
to its geographical neighborhood.
#' \item{\code{\link{bidev}}}: multiscalar typology based on 2 deviations.
\item{\code{\link{mst}}}: multiscalar typology based on the 3 deviations.
\item{\code{\link{mas}}}: multiscalar absolute synthesis, total amount of 
redistributions based on the three deviations. 
\item{\code{\link{map_bidev}}}: creating bidev and parameters for producing 
a map based on it. 
\item{\code{\link{map_mst}}}: creating mst and parameters for producing 
a map based on it.
\item{\code{\link{plot_bidev}}}: creating a plot for visualizing bidev results. 
\item{\code{\link{plot_mst}}}: creating a plot adapted for visualzing mst results.
}
}
\references{
GRASLAND C., YSEBAERT R., ZANIN C., LAMBERT N., Spatial disparities in Europe  (Chapter 4)  
in GLOERSEN E., DUBOIS A. (coord.), 2007, Regional disparities and cohesion: What Strategies for the future?, DG-IPOL – European Parliament.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/riatelab/MTA/}
  \item Report bugs at \url{https://github.com/riatelab/MTA/issues/}
}

}
\author{
\strong{Maintainer}: Timothée Giraud \email{timothee.giraud@cnrs.fr}

Authors:
\itemize{
  \item Ronan Ysebaert \email{ronan.ysebaert@cnrs.fr}
  \item Nicolas Lambert \email{nicolas.lambert@cnrs.fr}
}

}
