% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MTA}
\alias{MTA}
\alias{MTA-package}
\title{Multiscalar Territorial Analysis}
\description{
Build multiscalar territorial analysis based on various contexts. \cr
Main functions : 
\itemize{
\item{\code{\link{gdev}}: general deviation between regional ratios and a 
ratio of reference.}
\item{\code{\link{tdev}}: territorial deviation between regional ratios and 
ratios of an aggregated level.}
\item{\code{\link{sdev}}}: spatial deviation between regional ratios and ratios
of neigborhing regions.
\item{\code{\link{mst}}}: multiscalar typology based on the three deviations.
\item{\code{\link{mas}}}: multiscalar absolute synthesis, total amount of 
redistributions based on the three deviations. 
}
}
\references{
RONNNNNAAAAAANNNNN, A Great Paper Made With Love,
\emph{and in English}, 2023
}

