% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MTA}
\alias{MTA}
\alias{MTA-package}
\title{Multiscalar Territorial Analysis}
\description{
Build multiscalar territorial analysis based on various contexts for a given ratio defined by a numerator and a denominator.\cr
Main functions : 
\itemize{
\item{\code{\link{gdev}}: general deviation of each territorial unit as regards  
to all the study area (or a reference value).}
\item{\code{\link{tdev}}: territorial deviation of each territorial unit as regards  
to an intermediate territorial level of reference.}
\item{\code{\link{sdev}}}: spatial deviation of each territorial unit as regards  
to its geographical neighborhood.
\item{\code{\link{mst}}}: multiscalar typology based on the three deviations.
\item{\code{\link{mas}}}: multiscalar absolute synthesis, total amount of 
redistributions based on the three deviations. 
\item{\code{\link{mapmst}}}: multiscalar typology map. 
}
}
\references{
GRASLAND C., YSEBAERT R., ZANIN C., LAMBERT N., Spatial disparities in Europe  (Chapter 4)  
in GLOERSEN E., DUBOIS A. (coord.), 2007, Regional disparities and cohesion: What Strategies for the future?, DG-IPOL – European Parliament.
}
