% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{synthesis3}
\alias{synthesis3}
\title{Multiscalar typology}
\usage{
synthesis3(spdf, x, spdfid = NULL, xid = NULL, var1, var2, ref = NULL,
  key, order = NULL, dist = NULL, mat = NULL, threshold,
  superior = FALSE)
}
\arguments{
\item{spdf}{Spatial polygon data frame}

\item{x}{h}

\item{spdfid}{g}

\item{xid}{dhd}

\item{var1}{Numerator}

\item{var2}{Denominator}

\item{ref}{distance (in the units of the basemap) between centroids to determinate the local context. 0 = contiguity.}

\item{key}{Aggregation ky for the medium deviatiopn}

\item{order}{Threshold of typology}

\item{dist}{If true, the criterium is "gretter than". If false, the criterium is "lower than"}

\item{mat}{If true, the typology is ploted}

\item{threshold}{fh}

\item{superior}{hdfsh}
}
\description{
This function compute a multiscalar typology according to the three relative 
positions about contexts (global, medium, local). The elementary units are classified
 in eight classes according to their three relative positions.
}
\details{
xxx
}

